requirements.txt
fastapi==0.100.0
uvicorn==0.23.0
sqlalchemy==2.0.19
pydantic==2.0.3
pydantic[email]
pydantic-settings==2.0.3
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
bcrypt==4.0.1
alembic==1.11.1
asyncpg==0.28.0
python-dotenv==1.0.0
pytest==7.4.0
httpx==0.24.1
redis==4.6.0
celery==5.3.1
psycopg2-binary==2.9.6
pymongo==4.4.1
motor==3.2.0
emails==0.6
pyzbar==0.1.9
pillow==10.0.0
pyjwt==2.8.0
itsdangerous==2.1.2
psycopg2